Changed files

+++ 18.四数之和.cpp
@@ -26,7 +26,7 @@ public:
 
                 int left = j + 1, right = n - 1;
                 while (left < right) {
-                    long sum = nums[i] + nums[j] + nums[left] + nums[right];
+                    long sum = (long)nums[i] + nums[j] + nums[left] + nums[right];
                     if (sum == target) {
                         res.push_back({nums[i], nums[j], nums[left], nums[right]});
                         while (left < right && nums[left] == nums[left + 1]) {
